Updated some robot libraries
