+++ Libraries/Bigli.xml
@@ -1,0 +1,23 @@
+﻿<RobotSystems>
+    <RobotCell name="Bigli-KR20" manufacturer="KUKA">
+    <Mechanisms group="0">
+      <RobotArm model="KR20-3" manufacturer="KUKA" payload="20">
+        <Base x="0.000" y="0.000" z="0.000" q1="1.000" q2="0.000" q3="0.000" q4="0.000"/>
+        <Joints>
+          <Revolute number="1" a ="260" d ="675" minrange ="-185" maxrange ="185" maxspeed ="156"/>
+          <Revolute number="2" a ="680" d ="0" minrange ="-155" maxrange ="35" maxspeed ="156"/>
+          <Revolute number="3" a ="-35" d ="0" minrange ="-130" maxrange ="154" maxspeed ="156"/>
+          <Revolute number="4" a ="0" d ="670" minrange ="-350" maxrange ="350" maxspeed ="330"/>
+          <Revolute number="5" a ="0" d ="0" minrange ="-130" maxrange ="130" maxspeed ="332"/>
+          <Revolute number="6" a ="0" d ="158" minrange ="-350" maxrange ="350" maxspeed ="616"/>
+        </Joints>
+      </RobotArm>
+    </Mechanisms>
+    <IO>
+      <DO names="1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16"/>
+      <DI names="1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16"/>
+      <AO names="1,2,3,4"/>
+      <AI names="1,2,3,4"/>
+    </IO>
+  </RobotCell>
+</RobotSystems>
