+++ Libraries/iaac.xml
@@ -1,0 +1,25 @@
+﻿<RobotSystems>
+
+  <RobotCell name="iaac-KR150-2_110" manufacturer="KUKA">
+    <Mechanisms>
+      <RobotArm model="KR150-2_110" manufacturer="KUKA" payload="110">
+        <Base x="0.000" y="0.000" z="0.000" q1="1.000" q2="0.000" q3="0.000" q4="0.000"/>
+        <Joints>
+          <Revolute number="1" a ="350" d ="750" minrange = "-185" maxrange ="185" maxspeed ="105"/>
+          <Revolute number="2" a ="1250" d ="0" minrange = "-146" maxrange ="0" maxspeed ="110"/>
+          <Revolute number="3" a ="-55" d ="0" minrange = "-119" maxrange ="155" maxspeed ="93"/>
+          <Revolute number="4" a ="0" d ="1500" minrange = "-350" maxrange ="350" maxspeed ="170"/>
+          <Revolute number="5" a ="0" d ="0" minrange = "-125" maxrange ="125" maxspeed ="170"/>
+          <Revolute number="6" a ="0" d ="230" minrange = "-350" maxrange ="350" maxspeed ="238"/>
+        </Joints>
+      </RobotArm>
+    </Mechanisms>
+    <IO>
+      <DO names="1,2,3,4"/>
+      <DI names="1,2,3,4"/>
+      <AO names="1,2"/>
+      <AI names="1,2"/>
+    </IO>
+  </RobotCell>
+
+</RobotSystems>
