+++ Libraries/unimelb.xml
@@ -1,0 +1,43 @@
+﻿<RobotSystems>
+
+  <RobotCell name="UniMelb-IRB120" manufacturer="ABB">
+    <Mechanisms>
+      <RobotArm model="IRB120" manufacturer="ABB" payload="3">
+        <Base x="0.000" y="0.000" z="0.000" q1="1.000" q2="0.000" q3="0.000" q4="0.000"/>
+        <Joints>
+          <Revolute number="1" a ="0" d ="290" minrange = "-165" maxrange ="165" maxspeed ="250"/>
+          <Revolute number="2" a ="270" d ="0" minrange = "-110" maxrange ="110" maxspeed ="250"/>
+          <Revolute number="3" a ="70" d ="0" minrange = "-110" maxrange ="70" maxspeed ="250"/>
+          <Revolute number="4" a ="0" d ="302" minrange = "-160" maxrange ="160" maxspeed ="320"/>
+          <Revolute number="5" a ="0" d ="0" minrange = "-120" maxrange ="120" maxspeed ="320"/>
+          <Revolute number="6" a ="0" d ="72" minrange = "-400" maxrange ="400" maxspeed ="420"/>
+        </Joints>
+      </RobotArm>
+    </Mechanisms>
+    <IO>
+      <DO names="DO10_1,DO10_2,DO10_3,DO10_4,DO10_5,DO10_6,DO10_7,DO10_8,DO10_9,DO10_10,DO10_11,DO10_12,DO10_13,DO10_14,DO10_15,DO10_16"/>
+      <DI names="DI10_1,DI10_2,DI10_3,DI10_4,DI10_5,DI10_6,DI10_7,DI10_8,DI10_9,DI10_10,DI10_11,DI10_12,DI10_13,DI10_14,DI10_15,DO10_16"/>
+    </IO>
+  </RobotCell>
+
+  <RobotCell name="UniMelb-IRB1200" manufacturer="ABB">
+    <Mechanisms>
+      <RobotArm model="IRB1200" manufacturer="ABB" payload="5">
+        <Base x="0.000" y="0.000" z="0.000" q1="1.000" q2="0.000" q3="0.000" q4="0.000"/>
+        <Joints>
+          <Revolute number="1" a ="0" d ="399.1" minrange = "-170" maxrange ="170" maxspeed ="288"/>
+          <Revolute number="2" a ="448" d ="0" minrange = "-100" maxrange ="130" maxspeed ="240"/>
+          <Revolute number="3" a ="42" d ="0" minrange = "-200" maxrange ="70" maxspeed ="300"/>
+          <Revolute number="4" a ="0" d ="451" minrange = "-270" maxrange ="270" maxspeed ="400"/>
+          <Revolute number="5" a ="0" d ="0" minrange = "-130" maxrange ="130" maxspeed ="405"/>
+          <Revolute number="6" a ="0" d ="82" minrange = "-400" maxrange ="400" maxspeed ="600"/>
+        </Joints>
+      </RobotArm>
+    </Mechanisms>
+    <IO>
+      <DO names="DO10_1,DO10_2,DO10_3,DO10_4,DO10_5,DO10_6,DO10_7,DO10_8,DO10_9,DO10_10,DO10_11,DO10_12,DO10_13,DO10_14,DO10_15,DO10_16"/>
+      <DI names="DI10_1,DI10_2,DI10_3,DI10_4,DI10_5,DI10_6,DI10_7,DI10_8,DI10_9,DI10_10,DI10_11,DI10_12,DI10_13,DI10_14,DI10_15,DO10_16"/>
+    </IO>
+  </RobotCell>
+
+</RobotSystems>
