+++ RobotsGH/Kinematics.cs
@@ -16,7 +16,7 @@ namespace Robots.Grasshopper
         public Kinematics() : base("Kinematics", "K", "Inverse and forward kinematics for a single target (or group of targets in a robot cell with coord", "Robots", "Components") { }
         public override GH_Exposure Exposure => GH_Exposure.quarternary;
         public override Guid ComponentGuid => new Guid("{EFDA05EB-B281-4703-9C9E-B5F98A9B2E1D}");
-        protected override System.Drawing.Bitmap Icon => Properties.Resources.iconKinematics;
+        protected override Bitmap Icon => Properties.Resources.iconKinematics;
 
         protected override void RegisterInputParams(GH_InputParamManager pManager)
         {
