+++ RobotsGH/RobotsGH.csproj
@@ -10,7 +10,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Robots.Grasshopper</RootNamespace>
     <AssemblyName>RobotsGH</AssemblyName>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.7</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <IsWebBootstrapper>false</IsWebBootstrapper>
     <TargetFrameworkProfile />
@@ -73,7 +73,6 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Remote.cs" />
-    <Compile Include="RobotsInfo.cs" />
     <Compile Include="Target.cs" />
     <Compile Include="Util.cs" />
     <Compile Include="Visualization.cs" />
